+++ src/reanimated2/layoutReanimation/web/animationsManager.ts
@@ -9,7 +9,6 @@ import { createAnimationWithExistingTransform } from './createAnimation';
 import {
   extractTransformFromStyle,
   getProcessedConfig,
-  makeElementVisible,
   handleExitingAnimation,
   handleLayoutTransition,
   setElementAnimation,
@@ -17,6 +16,7 @@ import {
 import { areDOMRectsEqual } from './domUtils';
 import type { TransformsStyle } from 'react-native';
 import type { TransitionData } from './animationParser';
+import { makeElementVisible } from './componentStyle';
 
 function chooseConfig<ComponentProps extends Record<string, unknown>>(
   animationType: LayoutAnimationType,
