+++ src/reanimated2/layoutReanimation/web/componentUtils.ts
@@ -131,22 +131,6 @@ export function saveSnapshot(element: HTMLElement) {
   snapshots.set(element, element.getBoundingClientRect());
 }
 
-export function makeElementVisible(element: HTMLElement, delay: number) {
-  if (delay === 0) {
-    _updatePropsJS(
-      { visibility: 'initial' },
-      { _component: element as ReanimatedHTMLElement }
-    );
-  } else {
-    setTimeout(() => {
-      _updatePropsJS(
-        { visibility: 'initial' },
-        { _component: element as ReanimatedHTMLElement }
-      );
-    }, delay * 1000);
-  }
-}
-
 export function setElementAnimation(
   element: HTMLElement,
   animationConfig: AnimationConfig,
