+++ src/reanimated2/layoutReanimation/web/dummyPosition.ts
@@ -1,6 +1,25 @@
 'use strict';
 
+import { _updatePropsJS } from '../../js-reanimated';
+import type { ReanimatedHTMLElement } from '../../js-reanimated';
+
 export const snapshots = new WeakMap<HTMLElement, DOMRect>();
+
+export function makeElementVisible(element: HTMLElement, delay: number) {
+  if (delay === 0) {
+    _updatePropsJS(
+      { visibility: 'initial' },
+      { _component: element as ReanimatedHTMLElement }
+    );
+  } else {
+    setTimeout(() => {
+      _updatePropsJS(
+        { visibility: 'initial' },
+        { _component: element as ReanimatedHTMLElement }
+      );
+    }, delay * 1000);
+  }
+}
 
 function fixElementPosition(
   element: HTMLElement,
